Stylesheet copied to output dir first
Moved fit.css stylesheet handling from StoryTestFolder to
StoryTestFile. Stylesheet is always written out before any
results.

Added unit test to make sure that fit.css ends up in output
directory.
